Gracefully shutdown to SIGTERM
The stream client is a bit odd in how it handles async behavior. It
creates it's own cancel func internally and emits it on error. However,
we need to be able to cancel the context ourselves, so we create our own
cancelable context that is passed into the subscribe method. This will
need to be cleaned up.

Fixes #3